# src/backend/openapi/api-contracts-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EasyPassword API",
    "description": "Passwordless vault with WebAuthn and end-to-end encryption",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/register/options": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register Options",
        "description": "Initiate WebAuthn registration for a new device.\n\nThis endpoint generates WebAuthn registration options and stores a challenge\nin Redis for one-time consumption during verification. If the user doesn't exist,\nthey will be created. If the user already has an active device, a 409 Conflict\nis returned (single-device policy).\n\nReturns:\n    registration_id: Unique identifier for this registration session\n    public_key: WebAuthn registration options for navigator.credentials.create()",
        "operationId": "register_options_api_v1_auth_register_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationInitiationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationInitiationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Reauthentication required",
                  "code": "ReauthenticationRequired"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/options": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Options",
        "description": "Initiate WebAuthn authentication for an existing device.\n\nThis endpoint returns WebAuthn assertion options and stores a one-time\nchallenge in Redis. The login flow is only valid for an active device.",
        "operationId": "login_options_api_v1_auth_login_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationInitiationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationInitiationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Reauthentication required",
                  "code": "ReauthenticationRequired"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Verify",
        "description": "Complete WebAuthn authentication and issue session tokens.\n\nThe challenge is consumed one-time from Redis. After successful verification,\nan access token is returned and a refresh token cookie is set.",
        "operationId": "login_verify_api_v1_auth_login_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationCompletionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Reauthentication required",
                  "code": "ReauthenticationRequired"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register Verify",
        "description": "Complete WebAuthn registration and issue initial session.\n\nThis endpoint verifies the WebAuthn attestation credential, persists the device,\nand issues initial session tokens. The challenge is consumed one-time from Redis\nand cannot be reused. After successful verification, an access token is returned\nin the response and a refresh token is set in an HttpOnly cookie.\n\nReturns:\n    access_token: JWT access token (5-minute TTL)\n    device_id: UUID of the registered device\n    user_id: UUID of the user\n    token_type: Always \"Bearer\"",
        "operationId": "register_verify_api_v1_auth_register_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationCompletionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or reauthentication required.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Reauthentication required",
                  "code": "ReauthenticationRequired"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthenticationCompletionRequest": {
        "properties": {
          "authentication_id": {
            "type": "string",
            "title": "Authentication Id",
            "description": "The authentication_id returned from initiation endpoint"
          },
          "credential": {
            "additionalProperties": true,
            "type": "object",
            "title": "Credential",
            "description": "Full assertion payload from navigator.credentials.get()"
          }
        },
        "type": "object",
        "required": [
          "authentication_id",
          "credential"
        ],
        "title": "AuthenticationCompletionRequest",
        "description": "Request schema for WebAuthn authentication completion."
      },
      "AuthenticationCompletionResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "device_id",
          "user_id"
        ],
        "title": "AuthenticationCompletionResponse",
        "description": "Response schema for WebAuthn authentication completion."
      },
      "AuthenticationInitiationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "AuthenticationInitiationRequest",
        "description": "Request schema for WebAuthn authentication initiation."
      },
      "AuthenticationInitiationResponse": {
        "properties": {
          "authentication_id": {
            "type": "string",
            "title": "Authentication Id",
            "description": "Server-generated unique identifier for this authentication session"
          },
          "public_key": {
            "additionalProperties": true,
            "type": "object",
            "title": "Public Key",
            "description": "WebAuthn authentication options for navigator.credentials.get()"
          }
        },
        "type": "object",
        "required": [
          "authentication_id",
          "public_key"
        ],
        "title": "AuthenticationInitiationResponse",
        "description": "Response schema for WebAuthn authentication initiation."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RegistrationCompletionRequest": {
        "properties": {
          "registration_id": {
            "type": "string",
            "title": "Registration Id",
            "description": "The registration_id returned from initiation endpoint"
          },
          "credential": {
            "additionalProperties": true,
            "type": "object",
            "title": "Credential",
            "description": "Full attestation payload from navigator.credentials.create()"
          }
        },
        "type": "object",
        "required": [
          "registration_id",
          "credential"
        ],
        "title": "RegistrationCompletionRequest",
        "description": "Request schema for WebAuthn registration completion."
      },
      "RegistrationCompletionResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "device_id",
          "user_id"
        ],
        "title": "RegistrationCompletionResponse",
        "description": "Response schema for WebAuthn registration completion."
      },
      "RegistrationInitiationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name",
            "description": "User-friendly name for the device (e.g., 'iPhone', 'MacBook')"
          },
          "device_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Metadata",
            "description": "Optional device metadata (platform, browser, etc.)"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "RegistrationInitiationRequest",
        "description": "Request schema for WebAuthn registration initiation."
      },
      "RegistrationInitiationResponse": {
        "properties": {
          "registration_id": {
            "type": "string",
            "title": "Registration Id",
            "description": "Server-generated unique identifier for this registration session"
          },
          "public_key": {
            "additionalProperties": true,
            "type": "object",
            "title": "Public Key",
            "description": "WebAuthn registration options for navigator.credentials.create()"
          }
        },
        "type": "object",
        "required": [
          "registration_id",
          "public_key"
        ],
        "title": "RegistrationInitiationResponse",
        "description": "Response schema for WebAuthn registration initiation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}